refactor: moved manifest loading functions in CLI from project folder to manifest folder
