+++ packages/cli/src/lib/index.ts
@@ -12,3 +12,4 @@ export * from "./Compiler";
 export * from "./CodeGenerator";
 export * from "./SchemaComposer";
 export * from "./workflow";
+export * from "./manifest";
