+++ packages/cli/src/lib/manifest/index.ts
@@ -1,2 +1,3 @@
 export * from "./schema";
 export * from "./migrate";
+export * from "./load";
