+++ packages/cli/src/lib/manifest/load/loadInfraManifest.ts
@@ -1,23 +1,28 @@
-import { displayPath, withSpinner, intlMsg } from "../../../";
-
 import {
-  AppManifest,
-  deserializeAppManifest,
-} from "@polywrap/polywrap-manifest-types-js";
-import fs from "fs";
+  displayPath,
+  loadEnvironmentVariables,
+  withSpinner,
+  intlMsg,
+} from "../../";
 
-export const defaultAppManifest = [
-  "polywrap.yaml",
-  "polywrap.app.yaml",
-  "polywrap.app.yml",
+import fs from "fs";
+import {
+  InfraManifest,
+  deserializeInfraManifest,
+} from "@polywrap/polywrap-manifest-types-js";
+
+export const defaultInfraManifest = [
+  "polywrap.infra.yaml",
+  "polywrap.infra.yml",
 ];
 
-export async function loadAppManifest(
+export async function loadInfraManifest(
   manifestPath: string,
   quiet = false
-): Promise<AppManifest> {
-  const run = (): Promise<AppManifest> => {
+): Promise<InfraManifest> {
+  const run = (): Promise<InfraManifest> => {
     const manifest = fs.readFileSync(manifestPath, "utf-8");
+
     if (!manifest) {
       const noLoadMessage = intlMsg.lib_helpers_manifest_unableToLoad({
         path: `${manifestPath}`,
@@ -26,7 +31,10 @@ export async function loadAppManifest(
     }
 
     try {
-      const result = deserializeAppManifest(manifest);
+      let result = deserializeInfraManifest(manifest);
+      result = (loadEnvironmentVariables(
+        (result as unknown) as Record<string, unknown>
+      ) as unknown) as InfraManifest;
       return Promise.resolve(result);
     } catch (e) {
       return Promise.reject(e);
@@ -41,9 +49,9 @@ export async function loadAppManifest(
       intlMsg.lib_helpers_manifest_loadText({ path: manifestPath }),
       intlMsg.lib_helpers_manifest_loadError({ path: manifestPath }),
       intlMsg.lib_helpers_manifest_loadWarning({ path: manifestPath }),
-      async (_spinner) => {
+      async () => {
         return await run();
       }
-    )) as AppManifest;
+    )) as InfraManifest;
   }
 }
