+++ packages/cli/src/lib/manifest/load/loadPolywrapManifest.ts
@@ -1,0 +1,46 @@
+import { displayPath, withSpinner, intlMsg } from "../../";
+
+import {
+  PolywrapManifest,
+  deserializePolywrapManifest,
+} from "@polywrap/polywrap-manifest-types-js";
+import fs from "fs";
+
+export const defaultPolywrapManifest = ["polywrap.yaml", "polywrap.yml"];
+
+export async function loadPolywrapManifest(
+  manifestPath: string,
+  quiet = false
+): Promise<PolywrapManifest> {
+  const run = (): Promise<PolywrapManifest> => {
+    const manifest = fs.readFileSync(manifestPath, "utf-8");
+
+    if (!manifest) {
+      const noLoadMessage = intlMsg.lib_helpers_manifest_unableToLoad({
+        path: `${manifestPath}`,
+      });
+      throw Error(noLoadMessage);
+    }
+
+    try {
+      const result = deserializePolywrapManifest(manifest);
+      return Promise.resolve(result);
+    } catch (e) {
+      return Promise.reject(e);
+    }
+  };
+
+  if (quiet) {
+    return await run();
+  } else {
+    manifestPath = displayPath(manifestPath);
+    return (await withSpinner(
+      intlMsg.lib_helpers_manifest_loadText({ path: manifestPath }),
+      intlMsg.lib_helpers_manifest_loadError({ path: manifestPath }),
+      intlMsg.lib_helpers_manifest_loadWarning({ path: manifestPath }),
+      async () => {
+        return await run();
+      }
+    )) as PolywrapManifest;
+  }
+}
