+++ packages/cli/src/lib/option-defaults/defaultManifestFiles.ts
@@ -2,7 +2,7 @@ import {
   defaultAppManifest,
   defaultPluginManifest,
   defaultPolywrapManifest,
-} from "../project";
+} from "../manifest";
 
 const filterUniqueFn = (value: string, index: number, self: Array<string>) =>
   self.indexOf(value) === index;
