+++ packages/cli/src/lib/project/manifests/app/index.ts
@@ -1,2 +1,1 @@
 export * from "./languages";
-export * from "./load";
