+++ packages/cli/src/lib/project/manifests/plugin/index.ts
@@ -1,2 +1,1 @@
 export * from "./languages";
-export * from "./load";
